refactor: 전체적으로 리팩토링을 수행한다
