+++ src/components/@shared/ResultBox/ResultBox.styled.jsx
