+++ src/components/ProductDetail/ProductDetail.jsx
@@ -1,23 +1,28 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { PropTypes } from "prop-types";
-import { useProduct } from "../../hooks/useProduct";
 import * as S from "./ProductDetail.styles";
-import { formatPrice } from "../../utils/util";
-import Button from "../@shared/Button/Button";
+
+import { useProduct } from "../../hooks/useProduct";
 import { useCart } from "../../hooks/useCart";
+
+import { formatPrice } from "../../utils/util";
 import { CART } from "../../constants/constant";
+import Loading from "../@shared/Loading/Loading";
+import Button from "../@shared/Button/Button";
 
 const ProductDetail = ({ match }) => {
-  const { products, loading } = useProduct();
-
   const { items: cart, addCart, loading: cartLoading } = useCart();
-
+  const { products, loading: productLoading, getProducts } = useProduct();
   const product = products[match.params.id] ?? {
     name: "",
     thumbnail: "",
     price: 0,
   };
   const amount = cart[product?.id]?.amount ?? 0;
+
+  useEffect(() => {
+    getProducts();
+  }, []);
 
   const addToCart = () => {
     if (amount >= CART.MAX_AMOUNT) {
@@ -32,28 +37,23 @@ const ProductDetail = ({ match }) => {
 
   return (
     <S.ProductDetail>
-      {loading ? (
-        "상품 정보를 불러오고 있습니다"
-      ) : (
-        <>
-          <S.ImgWrapper>
-            <S.Img alt={product.name} src={product.thumbnail} />
-          </S.ImgWrapper>
-          <S.Name>{product.name}</S.Name>
-          <S.Price>{`${formatPrice(product.price)}원`}</S.Price>
-          <S.Cart>
-            {cartLoading ? (
-              <Button onClick={() => {}} disabled>
-                장바구니에 상품 추가중입니다
-              </Button>
-            ) : (
-              <Button onClick={addToCart}>
-                장바구니에 담기(현재 {amount}개)
-              </Button>
-            )}
-          </S.Cart>
-        </>
+      {(productLoading || cartLoading) && (
+        <Loading>상품목록을 불러오는 중입니다</Loading>
       )}
+      <S.ImgWrapper>
+        <S.Img alt={product.name} src={product.thumbnail} />
+      </S.ImgWrapper>
+      <S.Name>{product.name}</S.Name>
+      <S.Price>{`${formatPrice(product.price)}원`}</S.Price>
+      <S.Cart>
+        {cartLoading ? (
+          <Button onClick={() => {}} disabled>
+            장바구니에 상품 추가중입니다
+          </Button>
+        ) : (
+          <Button onClick={addToCart}>장바구니에 담기(현재 {amount}개)</Button>
+        )}
+      </S.Cart>
     </S.ProductDetail>
   );
 };
